move some file to specific folder and add rule marker

+++ Exploring Charts/model/DailySalesType.swift
@@ -3,7 +3,7 @@ import SwiftUI
 struct DailySalesType: Identifiable {
     let id = UUID()
     let day: String
-    let sales: Int
+    let sales: Double
 }
 
 let defaultDailySales: [DailySalesType] = [
